+++ src/device/types.h
@@ -81,4 +81,15 @@ union OsEvent
 	TextEvent text;
 };
 
+struct TimestepPolicy
+{
+	enum Enum
+	{
+		VARIABLE,
+		SMOOTHED,
+
+		COUNT
+	};
+};
+
 } // namespace crown