+++ docs/changelog.rst
@@ -9,7 +9,9 @@ Changelog
 
 **Runtime**
 
+* Added the ability to set a new (smoothed) timestep policy.
 * Improved ``graph`` command with the ability to add multiple fields, customize colors and limit the number of samples shown.
+* Lua API: Added ``Device.set_timestep_policy()`` and ``Device.set_timestep_smoothing()``.
 * Windows: fixed xinput.dll not found on some systems.
 * Windows: fixed console output and absolute paths when launched under MinGW.